+++ lib/api_guard/jwt_auth/blacklist_token.rb
@@ -2,33 +2,33 @@
 
 module ApiGuard
   module JwtAuth
-    # Common module for token blacklisting functionality
-    module BlacklistToken
-      def blacklisted_token_association(resource)
-        resource.class.blacklisted_token_association
+    # Common module for token revocation functionality
+    module RevokeToken
+      def revoked_token_association(resource)
+        resource.class.revoked_token_association
       end
 
-      def token_blacklisting_enabled?(resource)
-        blacklisted_token_association(resource).present?
+      def token_revocation_enabled?(resource)
+        revoked_token_association(resource).present?
       end
 
-      def blacklisted_tokens_for(resource)
-        blacklisted_token_association = blacklisted_token_association(resource)
-        resource.send(blacklisted_token_association)
+      def revoked_tokens_for(resource)
+        revoked_token_association = revoked_token_association(resource)
+        resource.send(revoked_token_association)
       end
 
-      # Returns whether the JWT token is blacklisted or not
-      def blacklisted?(resource)
-        return false unless token_blacklisting_enabled?(resource)
+      # Returns whether the JWT token is revoked or not
+      def revoked?(resource)
+        return false unless token_revocation_enabled?(resource)
 
-        blacklisted_tokens_for(resource).exists?(token: @token)
+        revoked_tokens_for(resource).exists?(token: @token)
       end
 
-      # Blacklist the current JWT token from future access
-      def blacklist_token
-        return unless token_blacklisting_enabled?(current_resource)
+      # Revoke the current JWT token from future access
+      def revoke_token
+        return unless token_revocation_enabled?(current_resource)
 
-        blacklisted_tokens_for(current_resource).create(token: @token, expire_at: Time.at(@decoded_token[:exp]).utc)
+        revoked_tokens_for(current_resource).create(token: @token, expire_at: Time.at(@decoded_token[:exp]).utc)
       end
     end
   end