+++ lib/generators/api_guard/controllers/templates/passwords_controller.rb
@@ -6,7 +6,7 @@ module <%= @controller_scope %>
     #   invalidate_old_jwt_tokens(current_resource)
     #
     #   if current_resource.update_attributes(password_params)
-    #     blacklist_token unless ApiGuard.invalidate_old_tokens_on_password_change
+    #     revoke_token unless ApiGuard.invalidate_old_tokens_on_password_change
     #     destroy_all_refresh_tokens(current_resource)
     #
     #     create_token_and_set_header(current_resource, resource_name)