+++ spec/dummy/app/models/blacklisted_token.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class BlacklistedToken < ApplicationRecord
+class RevokedToken < ApplicationRecord
   belongs_to :user, optional: true
   belongs_to :admin, optional: true
 end