+++ spec/dummy/app/models/user.rb
@@ -3,7 +3,7 @@
 class User < ApplicationRecord
   has_secure_password
 
-  api_guard_associations refresh_token: 'refresh_tokens', blacklisted_token: 'blacklisted_tokens'
+  api_guard_associations refresh_token: 'refresh_tokens', revoked_token: 'revoked_tokens'
 
   # == Validations =====================================================================================================
   validates :email, presence: true
@@ -11,6 +11,6 @@ class User < ApplicationRecord
 
   # == Relationships ===================================================================================================
   has_many :refresh_tokens, dependent: :delete_all
-  has_many :blacklisted_tokens, dependent: :delete_all
+  has_many :revoked_tokens, dependent: :delete_all
   has_many :posts, dependent: :delete_all
 end